fix(broker): doubles forgotten

+++ broker/storage/test/perfdata.cc
@@ -243,9 +243,9 @@ TEST_F(StorageParserParsePerfdata, Simple2) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(18.0);
   expected.unit("%");
-  expected.warning(std::numeric_limits<double>::infinity());
+  expected.warning(std::numeric_limits<float>::infinity());
   expected.warning_low(15.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(10.0);
   expected.min(0.0);
   expected.max(100.0);
@@ -270,7 +270,7 @@ TEST_F(StorageParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(2.45698);
   expected.unit("s");
-  expected.max(std::numeric_limits<double>::infinity());
+  expected.max(std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ++it;
 
@@ -282,7 +282,7 @@ TEST_F(StorageParserParsePerfdata, Complex1) {
   expected.unit("B/s");
   expected.warning(5.0);
   expected.warning_low(0.0);
-  expected.min(-std::numeric_limits<double>::infinity());
+  expected.min(-std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ++it;
 
@@ -302,7 +302,7 @@ TEST_F(StorageParserParsePerfdata, Complex1) {
   expected.value(1234.0);
   expected.warning(10.0);
   expected.warning_low(0.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(11.0);
   ASSERT_TRUE(expected == *it);
   ++it;
@@ -313,7 +313,7 @@ TEST_F(StorageParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::counter);
   expected.value(1234.0);
   expected.warning(10.0);
-  expected.warning_low(-std::numeric_limits<double>::infinity());
+  expected.warning_low(-std::numeric_limits<float>::infinity());
   expected.critical(30.0);
   expected.critical_low(20.0);
   ASSERT_TRUE(expected == *it);
@@ -336,10 +336,10 @@ TEST_F(StorageParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(9.0);
   expected.unit("queries_per_second");
-  expected.warning(std::numeric_limits<double>::infinity());
+  expected.warning(std::numeric_limits<float>::infinity());
   expected.warning_low(10.0);
   expected.warning_mode(true);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(5.0);
   expected.critical_mode(true);
   expected.min(0.0);
@@ -454,7 +454,7 @@ TEST_F(StorageParserParsePerfdata, Complex2) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(2.45698);
   expected.unit("s");
-  expected.max(std::numeric_limits<double>::infinity());
+  expected.max(std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
   ++it;
@@ -467,7 +467,7 @@ TEST_F(StorageParserParsePerfdata, Complex2) {
   expected.unit("B/s");
   expected.warning(5.0);
   expected.warning_low(0.0);
-  expected.min(-std::numeric_limits<double>::infinity());
+  expected.min(-std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
   ++it;
@@ -499,7 +499,7 @@ TEST_F(StorageParserParsePerfdata, Complex2) {
   expected.value(1234.17);
   expected.warning(10.0);
   expected.warning_low(0.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(11.0);
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
@@ -511,7 +511,7 @@ TEST_F(StorageParserParsePerfdata, Complex2) {
   expected.value_type(misc::perfdata::counter);
   expected.value(1234.147);
   expected.warning(10.0);
-  expected.warning_low(-std::numeric_limits<double>::infinity());
+  expected.warning_low(-std::numeric_limits<float>::infinity());
   expected.critical(30.0);
   expected.critical_low(20.0);
   ASSERT_TRUE(expected == *it);
@@ -553,7 +553,7 @@ TEST_F(StorageParserParsePerfdata, BadMetric) {
   int i = 1;
   for (auto& p : lst) {
     ASSERT_EQ(p.name(), fmt::format("user{}", i));
-    ASSERT_EQ(p.value(), static_cast<double>(i));
+    ASSERT_EQ(p.value(), static_cast<float>(i));
     ++i;
   }
 }
@@ -566,7 +566,7 @@ TEST_F(StorageParserParsePerfdata, BadMetric1) {
   int i = 1;
   for (auto& p : lst) {
     ASSERT_EQ(p.name(), fmt::format("user{}", i));
-    ASSERT_EQ(p.value(), static_cast<double>(i));
+    ASSERT_EQ(p.value(), static_cast<float>(i));
     ++i;
   }
 }
